refactor: Rename `…/event-detection/event-types/alexaSkill.js`
